fix: seperated server and index so that jest wouldn't run the server
